requirements.txt

requests
beautifulsoup4
pandas
openpyxl
anthropic
python-dotenv
trafilatura